Release: Version 0.1.2 - Major Refactor, fixes known issues
