+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "obsidian-lean-journal",
-	"version": "0.0.1",
+	"version": "0.1.2",
 	"description": "This is a simple lean journal plugin for Obsidian (https://obsidian.md)",
 	"main": "main.js",
 	"scripts": {
@@ -19,6 +19,7 @@
 		"esbuild": "0.17.3",
 		"obsidian": "latest",
 		"prettier": "^3.3.3",
+		"process": "^0.11.10",
 		"tslib": "2.4.0",
 		"typescript": "4.7.4"
 	}
