+++ src/constants.ts
@@ -1,0 +1,40 @@
+export const JOURNAL_ICON = `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M4 19.5A2.5 2.5 0 0 1 6.5 17H20"></path><path d="M6.5 2H20v20H6.5A2.5 2.5 0 0 1 4 19.5v-15A2.5 2.5 0 0 1 6.5 2z"></path><line x1="12" y1="6" x2="18" y2="6"></line><line x1="12" y1="10" x2="18" y2="10"></line><line x1="12" y1="14" x2="18" y2="14"></line><line x1="12" y1="18" x2="18" y2="18"></line></svg>`;
+
+export const testLogs = [
+  `---
+  log:
+  - "[[2024-08-15]]"
+  - "[[2024-08-15]]"
+  - "[[2024-08-15]]"
+  - "[[2024-08-15]]"
+  ---
+  Content here`,
+
+  `---
+  log: 2024-01-01
+  log: 2024-01-01
+  log: 2024-01-01
+  log: 2024-01-01
+  ---
+  Content here`,
+
+  `---
+  log: "[[2024-01-01]]"
+  log: "[[2024-01-01]]"
+  ---
+  Content here`,
+
+  `---
+  title: Test
+  log:
+  - "[[2024-08-15]]"
+  date: 2024-08-15
+  ---
+  Content here`,
+
+  `---
+  log:
+  - "[[2024-08-19]]"
+  ---
+  Content here`
+];
