+++ tsconfig.json
@@ -11,14 +11,14 @@
     "importHelpers": true,
     "isolatedModules": true,
     "strictNullChecks": true,
-    "lib": [
-      "DOM",
-      "ES5",
-      "ES6",
-      "ES7"
-    ]
+	  "lib": [
+		  "DOM",
+		  "ES5",
+		  "ES6",
+		  "ES7"
+	  ]
   },
-  "include": [
-    "**/*.ts"
-  ]
+	"include": [
+		"**/*.ts"
+	]
 }
