+++ versions.json
@@ -1,3 +1,5 @@
 {
-  "0.1.0": "0.15.0"
+	"0.1.0": "0.15.0",
+	"0.1.1": "0.15.0",
+	"0.1.2": "0.15.0"
 }
